libstagefright: omx: Prevent assertion due to state mismatches
Instead handle them a bit better. Since the assertions dealt
with in this patchset can only happen if the OMX IL client died
while de-init'ing the component, it is impossible to deal with
all possible cases gracefully. Hence take an aggressive approach
by moving state of the component to OMX_StateInvalid. This looks
okay as per the spec, but might not be the best approach.

CRs-Fixed: 1077558
Change-Id: Id240a23f2c16e89f41f9156277f4dcbe0969f954